chore: move DELETE /api/price-comparisons/all to /api/price-comparisons
