+++ src/routes/api/price-comparisons/+server.ts
@@ -77,3 +77,14 @@ export async function POST({ request }): Promise<Response> {
     trackingId: resultId,
   });
 }
+
+export async function DELETE(): Promise<Response> {
+  const productRepo = await $productRepository;
+
+  await productRepo.removeAllComparisons();
+  await productRepo.removeAllHistory();
+
+  return new Response(null, {
+    status: 204,
+  });
+}
